Reorganize source by concern: split 1063-line engine, drop stale typed.go/cache_ prefixes, fix CLAUDE.md drift
