+++ emit_test.go
