+++ write_test.go
