+++ writelock.go
